Moved all web services into the assets/web-services directory.
