+++ assets/src/scripts/web-services/statistics-data.js
@@ -1,6 +1,6 @@
-import { get } from './ajax';
-import config from './config';
-import { parseRDB } from './utils';
+import { get } from '../ajax';
+import config from '../config';
+import { parseRDB } from '../utils';
