+++ assets/src/scripts/web-services/statistics-data.spec.js
